+++ helpers/global_helpers.ts
