+++ src/components/Comments/Comments.tsx
@@ -1,5 +1,5 @@
 import { useEditor } from "@tiptap/react";
-import { CommentEditorExtensions } from "../../../helpers/global-helpers";
+import { CommentEditorExtensions } from "../../../helpers/global_helpers";
 import CommentEditor from "../Editor/CommentEditor";
 import { useMemo, useState, useEffect } from "react";
 import CommentItem from "./CommentItem";
